+++ src/EnrolmentWorkflow/Action/SendNotification/SendNotificationRunner.php
@@ -3,6 +3,7 @@
 namespace srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Action\SendNotification;
 
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Action\AbstractActionRunner;
+use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Deputy\Deputy;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Request\Request;
 
 /**
@@ -52,6 +53,18 @@ class SendNotificationRunner extends AbstractActionRunner
 
             case SendNotification::TO_TYPE_SPECIFIC_USERS:
                 $users = $this->action->getToSpecificUsers();
+                break;
+
+            case SendNotification::TO_TYPE_RESPONSIBLE_USERS_AND_DEPUTIES:
+                $users = array_unique(array_reduce($request->getResponsibleUsers(), function (array $users, int $user_id) : array {
+                    $users[] = $user_id;
+
+                    $users = array_merge($users, array_map(function (Deputy $deputy) : int {
+                        return $deputy->getDeputyUserId();
+                    }, self::srUserEnrolment()->enrolmentWorkflow()->deputies()->getUserDeputies($user_id)));
+
+                    return $users;
+                }, []));
                 break;
 
             default: