+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [x]
 - Requests table: All, own, open
+- Send notification option to responsible users and deputies
 
 ## [3.3.0]
 - Improve user selection in assistants and deputies